Create helper functions jsonld/sparql.ts
